// src/render/utils/enum.ts
export enum CoreMouseMode {
    SELECTION = 'selection',
    FRAME = 'frame',
    TEXT = 'text',
}

export enum ObjectType {
    GROUP = 'group',
    FRAME = 'frame',
    TEXT = 'text',
    ACTIVE_SELECTION = 'activeselection',
}

export enum ElementName {
    FRAME = 'frame',
    TEXT = 'text',
    IMAGE = 'image',
    GROUP = 'group',
    LOADING_GROUP = 'loadingGroup',
    LOADING_RECT = 'loadingRect',
    LOADING_TEXT = 'loadingText',
}